packages/snomed-neo4j-core/pyproject.toml
[project]
name = "snomed-neo4j-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Bence Szabó", email = "bszabo96@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "neo4j>=5.28.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[project.scripts]
download = "snomed_neo4j_core.download:main"
load = "snomed_neo4j_core.loader:main"
slim = "snomed_neo4j_core.slim:main"
